% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gpkg.R
\name{create_gpkg_version_control}
\alias{create_gpkg_version_control}
\title{Title}
\usage{
create_gpkg_version_control(
  .data,
  dsn,
  layer,
  status = "valid",
  first_appearance = NULL,
  index_field = "Beteckn",
  index_name = "idx_avverkanm_beteckn",
  append = FALSE,
  log_table_name = "log_data",
  unlink = FALSE
)
}
\arguments{
\item{unlink}{}
}
\description{
Title
}
