% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gpkg.R
\name{create_initial_data}
\alias{create_initial_data}
\title{Title}
\usage{
create_initial_data(
  .data,
  dsn,
  layer,
  status = "valid",
  first_appearance = NULL,
  append = FALSE
)
}
\arguments{
\item{append}{}
}
\description{
Title
}
