% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data_forestry.R
\name{execute_db_updates}
\alias{execute_db_updates}
\title{Title}
\usage{
execute_db_updates(
  original,
  revised,
  query = NA,
  quiet = FALSE,
  base_dir = "downloads",
  dsn_out,
  layer_out,
  compare_attributes =
    "Lannr;Lan;Kommunnr;Kommun;ArendeAr;Beteckn;Avverktyp;Skogstyp;Inkomdatum;AnmaldHa;SkogsodlHa;NatforHa;ArendeStat;AvvSasong;AvvHa;Avverkning",
  match_type = 1,
  unchanged = qgisprocess::qgis_tmp_vector(),
  added = qgisprocess::qgis_tmp_vector(),
  deleted = qgisprocess::qgis_tmp_vector(),
  log_table_name = "log_data"
)
}
\arguments{
\item{log_table_name}{}
}
\description{
Title
}
