% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_updates.R
\name{execute_db_updates_memory}
\alias{execute_db_updates_memory}
\title{Title}
\usage{
execute_db_updates_memory(
  original,
  revised,
  original_str,
  revised_str,
  dsn_out,
  layer_out,
  compare_attributes = c("Lannr", "Lan", "Kommunnr", "Kommun", "ArendeAr", "Beteckn",
    "Avverktyp", "Skogstyp", "Inkomdatum", "AnmaldHa", "SkogsodlHa", "NatforHa",
    "ArendeStatus", "AvvSasong", "AvvHa", "AvverkningsanmalanKlass"),
  match_type = 1,
  unchanged = qgisprocess::qgis_tmp_vector(),
  added = qgisprocess::qgis_tmp_vector(),
  deleted = qgisprocess::qgis_tmp_vector(),
  log_table_name = "log_data"
)
}
\arguments{
\item{log_table_name}{}
}
\description{
Title
}
