% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data_forestry.R
\name{execute_db_updates_n}
\alias{execute_db_updates_n}
\title{Title}
\usage{
execute_db_updates_n(
  .data,
  dsn_out,
  layer_out,
  ...,
  n = base::seq_len(nrow(.data))
)
}
\arguments{
\item{n}{}
}
\description{
Title
}
