% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_updates.R
\name{list_files_compare}
\alias{list_files_compare}
\title{Title}
\usage{
list_files_compare(
  base_dir = "./downloads",
  file_name_pattern = "^sksAvverkAnm",
  date_pattern = "([0-9]{8})_?([0-9]{6})?",
  date_formats = c("ymd", "ymd_HMS"),
  tz = "Europe/Stockholm",
  drop_na = TRUE
)
}
\arguments{
\item{drop_na}{}
}
\description{
Title
}
