% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_avverk_beteckn}
\alias{pad_avverk_beteckn}
\title{Title}
\usage{
pad_avverk_beteckn(
  Beteckn,
  pattern = " ([0-9]{1,5})\\\\-",
  pad = "0",
  width = 5
)
}
\arguments{
\item{width}{}
}
\description{
Title
}
