% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sks_import}
\alias{sks_import}
\title{Title}
\usage{
sks_import(
  input,
  output,
  driver = "GPKG",
  assign_srs = "EPSG:3006",
  new_layer_name,
  new_layer_type = "PROMOTE_TO_MULTI"
)
}
\arguments{
\item{new_layer_type}{}
}
\description{
Title
}
