% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sks_last_updated}
\alias{sks_last_updated}
\title{Title}
\usage{
sks_last_updated(
  feed_name,
  base_url = "https://geodpags.skogsstyrelsen.se/geodataport",
  datetime_format = "\%Y-\%m-\%dT\%H:\%M:\%SZ",
  string_format = "\%Y\%m\%d_\%H\%M\%S"
)
}
\arguments{
\item{string_format}{}
}
\description{
Title
}
