% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data_forestry.R
\name{sks_log_table}
\alias{sks_log_table}
\title{Title}
\usage{
sks_log_table(
  dsn = "sks_avverkning_anmald_sverige.gpkg",
  table_name = "log_data"
)
}
\arguments{
\item{table_name}{}
}
\description{
Title
}
