% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{tmp_export_from_gpkg}
\alias{tmp_export_from_gpkg}
\title{Title}
\usage{
tmp_export_from_gpkg(
  layer_suffix,
  dsn = "sks_avverkning_anmald",
  layer_prefix = "sks_avverkning_anmald",
  download_dir = "./downloads",
  output_file = "sksAvverkAnm"
)
}
\arguments{
\item{output_file}{}
}
\description{
Title
}
