% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_updates.R
\name{find_pending_db_updates}
\alias{find_pending_db_updates}
\title{Title}
\usage{
find_pending_db_updates(start_date, dsn_out, log_table_name = "log_data", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
