% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_data_forestry.R
\name{list_files_compare}
\alias{list_files_compare}
\title{Title}
\usage{
list_files_compare(
  base_dir = "./downloads",
  prefix = "sksAvverkAnm",
  match_pattern = glue::glue("{prefix}\\\\_([0-9]{{8}})_?([0-9]{{6}})?.zip"),
  drop_na = TRUE
)
}
\arguments{
\item{drop_na}{}
}
\description{
Title
}
